+++ real-world-vue-learning/src/store/modules/user.js
@@ -1,0 +1,6 @@
+export const state = {
+  user: {
+    id: 'abc123',
+    name: 'Kelvin', // I removed the last name Jahr here our title is on one line
+  },
+}
