+++ real-world-vue-learning/src/views/EventShow.vue
@@ -26,13 +26,16 @@
   </div>
 </template>
 <script>
-import { mapState } from 'vuex'
+import { mapState, mapActions } from 'vuex'
 export default {
   props: ['id'],
   created() {
-    this.$store.dispatch('fetchEvent', this.id)
+    this.fetchEvent(this.id)
   },
-  computed: mapState(['event']),
+  computed: mapState({
+    event: state => state.event.event,
+  }),
+  methods: mapActions('event', ['fetchEvent']),
 }
 </script>
 <style scoped>
